+++ src/inspect_ai/_cli/info.py
@@ -5,7 +5,7 @@ import click
 from inspect_ai import __version__
 from inspect_ai._util.constants import PKG_PATH
 from inspect_ai._view.server import resolve_header_only
-from inspect_ai.log._file import eval_log_json, read_eval_log
+from inspect_ai.log._file import eval_log_json_str, read_eval_log
 
 from .log import headers, schema, types
 
@@ -46,7 +46,7 @@ def log(path: str, header_only: int) -> None:
     header_only = resolve_header_only(path, header_only)
 
     log = read_eval_log(path, header_only=header_only)
-    print(eval_log_json(log))
+    print(eval_log_json_str(log))
 
 
 @info_command.command("log-file-headers", hidden=True)