+++ src/inspect_ai/_eval/task/run.py
@@ -42,7 +42,7 @@ from inspect_ai.log import (
     EvalStats,
 )
 from inspect_ai.log._condense import condense_sample
-from inspect_ai.log._file import eval_log_json
+from inspect_ai.log._file import eval_log_json_str
 from inspect_ai.log._log import EvalSampleLimit, EvalSampleReductions, eval_error
 from inspect_ai.log._samples import active_sample
 from inspect_ai.log._transcript import (
@@ -217,7 +217,7 @@ async def task_run(options: TaskRunOptions) -> EvalLog:
         with display().task(profile) as td:
             try:
                 # start the log
-                log_start(logger, plan, generate_config)
+                await log_start(logger, plan, generate_config)
 
                 with td.progress() as p:
                     # forward progress
@@ -370,18 +370,20 @@ async def task_run(options: TaskRunOptions) -> EvalLog:
 
         # log as appropriate
         if cancelled:
-            eval_log = logger.log_finish("cancelled", stats, results, reductions)
+            eval_log = await logger.log_finish("cancelled", stats, results, reductions)
         elif error:
-            eval_log = logger.log_finish("error", stats, results, reductions, error)
+            eval_log = await logger.log_finish(
+                "error", stats, results, reductions, error
+            )
         else:
-            eval_log = logger.log_finish("success", stats, results, reductions)
+            eval_log = await logger.log_finish("success", stats, results, reductions)
 
         # notify the view module that an eval just completed
         # (in case we have a view polling for new evals)
         view_notify_eval(logger.location)
 
         try:
-            await send_telemetry("eval_log", eval_log_json(eval_log))
+            await send_telemetry("eval_log", eval_log_json_str(eval_log))
         except Exception as ex:
             py_logger.warning(
                 f"Error occurred sending telemetry: {exception_message(ex)}"
@@ -475,7 +477,7 @@ async def task_run_sample(
 
             # log if requested
             if logger:
-                logger.log_sample(previous_sample, flush=False)
+                await logger.log_sample(previous_sample, flush=False)
 
             # return score
             if previous_sample.scores:
@@ -674,7 +676,7 @@ async def task_run_sample(
                 state = state_without_base64_images(state)
 
             # log the sample
-            log_sample(
+            await log_sample(
                 logger=logger,
                 sample=sample,
                 state=state,
@@ -692,7 +694,7 @@ async def task_run_sample(
             return None
 
 
-def log_sample(
+async def log_sample(
     logger: TaskLogger,
     sample: Sample,
     state: TaskState,
@@ -738,7 +740,7 @@ def log_sample(
         limit=limit,
     )
 
-    logger.log_sample(condense_sample(eval_sample, log_images), flush=True)
+    await logger.log_sample(condense_sample(eval_sample, log_images), flush=True)
 
 
 async def resolve_dataset(